backend/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
transformers==4.57.3
torch==2.9.1
keybert==0.9.0
scikit-learn==1.8.0
supabase==2.25.0
python-jose==3.5.0
psycopg2-binary==2.9.11
pytest==9.0.2
requests==2.32.5

# Optional but useful
sentence-transformers==5.2.0
cryptography==46.0.3
httpx==0.28.1